consolidate ticket
